update: Add support for ext lib 16 (quickdesh/animiru#104)
* Implement ext lib 16-rc3

* Implement season migration

* Small cleanup

* Implement `16-rc4`

* ext lib 16 parameter addition fixes

* chore: changelog

* Fix name in migration

* Fix proguard rules for extensions

---
* Implement seasons feature (aniyomiorg/aniyomi##2095)
(cherry picked from commit b8b48b0296c921604f5f821389de44a45ed9272f)

* Fixes from 16-rc1 (aniyomiorg/aniyomi##2142)
(cherry picked from commit 3d9f3bf82d7428718660706e6cc28427cb4e41f8)

* Fix restoring backup (aniyomiorg/aniyomi##2148)
(cherry picked from commit 10d7ba14723a243341797c31b5f999b5716e81eb)

* fix: Remove `FetchType.Unknown` (aniyomiorg/aniyomi##2149)
(cherry picked from commit a14bdf1efb438ad6887e37020c8eede287b9b6e8)

* Fix issues regarding seasons (aniyomiorg/aniyomi##2154)
(cherry picked from commit 69779c9d8a659318a9278ffbfab4daf7ccc7da88

* feat: Introduce new parameters for ext lib 16 (aniyomiorg/aniyomi##1982)
(cherry picked from commit 06aaec0abd5d05956a32d8140bcde1e8f79aa947)

* Fix `getSeasonList` being called for non-season entries (aniyomiorg/aniyomi##2157)
(cherry picked from commit 2bab8d93e74efa6547116a708553ae9512acd7a7)

* Fix sql crash on migration (aniyomiorg/aniyomi##2159)
(cherry picked from commit 411d080d77bc328d29fa4fa81ad1a2dd89cc9f38)

* fix: ext lib 16 parameter addition fixes (aniyomiorg/aniyomi#2160)
(cherry picked from commit 191fea8385bea2112f071c2db5a682662cbd0e44)

---

(cherry picked from commit a29ce96b93578607682ee620dc43d45d84211565)

+++ app/src/main/java/mihon/domain/migration/usecases/MigrateMangaUseCase.kt
@@ -1,10 +1,14 @@
 package mihon.domain.migration.usecases
 
+import eu.kanade.domain.anime.interactor.SyncSeasonsWithSource
 import eu.kanade.domain.chapter.interactor.SyncChaptersWithSource
 import eu.kanade.domain.manga.interactor.UpdateManga
+import eu.kanade.domain.manga.model.hasCustomBackground
 import eu.kanade.domain.manga.model.hasCustomCover
 import eu.kanade.domain.manga.model.toSManga
 import eu.kanade.domain.source.service.SourcePreferences
+import eu.kanade.tachiyomi.animesource.model.FetchType
+import eu.kanade.tachiyomi.data.cache.BackgroundCache
 import eu.kanade.tachiyomi.data.cache.CoverCache
 import eu.kanade.tachiyomi.data.download.DownloadManager
 import eu.kanade.tachiyomi.data.track.EnhancedTracker
@@ -35,12 +39,18 @@ class MigrateMangaUseCase(
     private val updateManga: UpdateManga,
     private val getChaptersByMangaId: GetChaptersByMangaId,
     private val syncChaptersWithSource: SyncChaptersWithSource,
+    // AY -->
+    private val syncSeasonsWithSource: SyncSeasonsWithSource,
+    // <-- AY
     private val updateChapter: UpdateChapter,
     private val getCategories: GetCategories,
     private val setMangaCategories: SetMangaCategories,
     private val getTracks: GetTracks,
     private val insertTrack: InsertTrack,
     private val coverCache: CoverCache,
+    // AY -->
+    private val backgroundCache: BackgroundCache,
+    // <-- AY
     // KMK -->
     private val getHistory: GetHistory,
     private val upsertHistory: UpsertHistory,
@@ -61,16 +71,33 @@ class MigrateMangaUseCase(
         val flags = /* KMK --> */ presetFlags ?: /* KMK <-- */ sourcePreferences.migrationFlags().get()
 
         try {
-            val chapters = targetSource.getChapterList(target.toSManga())
+            // AY -->
+            when (target.fetchType) {
+                FetchType.Seasons -> {
+                    val seasons = targetSource.getSeasonList(target.toSManga())
 
-            try {
-                syncChaptersWithSource.await(chapters, target, targetSource)
-            } catch (_: Exception) {
-                // Worst case, chapters won't be synced
+                    try {
+                        syncSeasonsWithSource.await(seasons, target, targetSource)
+                    } catch (_: Exception) {
+                        // Worst case, seasons won't be synced
+                    }
+                }
+                FetchType.Episodes -> {
+                    // <-- AY
+                    val chapters = targetSource.getChapterList(target.toSManga())
+
+                    try {
+                        syncChaptersWithSource.await(chapters, target, targetSource)
+                    } catch (_: Exception) {
+                        // Worst case, chapters won't be synced
+                    }
+                }
             }
 
             // Update chapters read, bookmark and dateFetch
-            if (MigrationFlag.CHAPTER in flags) {
+            // AY -->
+            if (MigrationFlag.CHAPTER in flags && target.fetchType == FetchType.Episodes) {
+                // <-- AY
                 val prevMangaChapters = getChaptersByMangaId.await(current.id)
                 val mangaChapters = getChaptersByMangaId.await(target.id)
 
@@ -157,7 +184,11 @@ class MigrateMangaUseCase(
             }
 
             // Delete downloaded
-            if (MigrationFlag.REMOVE_DOWNLOAD in flags && currentSource != null) {
+            if (MigrationFlag.REMOVE_DOWNLOAD in flags && currentSource != null &&
+                // AY -->
+                current.fetchType == FetchType.Episodes
+                // <-- AY
+            ) {
                 downloadManager.deleteManga(current, currentSource)
             }
 
@@ -165,6 +196,16 @@ class MigrateMangaUseCase(
             if (MigrationFlag.CUSTOM_COVER in flags && current.hasCustomCover()) {
                 coverCache.setCustomCoverToCache(target, coverCache.getCustomCoverFile(current.id).inputStream())
             }
+
+            // AY -->
+            // Update custom background (recheck if custom background exists)
+            if (MigrationFlag.CUSTOM_BACKGROUND in flags && current.hasCustomBackground()) {
+                backgroundCache.setCustomBackgroundToCache(
+                    target,
+                    backgroundCache.getCustomBackgroundFile(current.id).inputStream(),
+                )
+            }
+            // <-- AY
 
             val currentMangaUpdate = MangaUpdate(
                 id = current.id,
